+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   - Fix TestName 
   - Fix direct execution of tests when the script is dot-sourced to global scope [GH-144]
   - Fix mock parameter filter in strict mode [GH-143]
+  - Fix nUnit schema compatibility
+  - Fix special characters in nUnit output
 
 ## 3.0.0-beta (June 24, 2014)
   - Add full support for module mocking